[Date/Person] Changed and add suggestions
Class Person:
-Implemented constructor, gets and sets functions;

Derivated Class Student:
-Inside of cicle while in constructor, maybe we should have a vector of
units per year to run and ask inside this loop;
In Student constructor such conditions should not be within the class
After verifying that there is a current vector students? Because so do
not give to see if the code exists, then my suggestion would be to make
these checks all in AddStudent (eg.) And create the student already in
the constructor pass arguments all including vectors.
The same happens to have a vector of curricular units per year to be
easier to check in the class course.
-Implemented gets and sets functions;

Derivated Class Teacher:
-Constructor:
Same as student, it will advantageous ask in class course to send vector
directly;
-Implemented sets and gets functions;
Class Date:
Implemented constructor, sets, gets and operators except o cin;