[Various] Big developments!
UTILS
- Corrected TAB and BIG_TAB macros (who the fuck put them with a number
instead of said number of spaces? .-.)

PERSON
- Correct small syntax errors
- Implemented operator<< overload for Student and updated the same one
for Teacher

MENUS
- Added cin.ignore before getline (will have to do for all getlines, but
this was the first listing I tested so it's only here for now because I
wanted to see if the data was OK and listings worked)

COURSE
- Updated reading loops for teachers and students according to new file
definition (tested and reading correctly)

Program ran for the first time, and it works! (I think)